requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Assets==2.1.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
logging==0.4.9.6
MarkupSafe==2.1.5
multidict==6.0.5
priority==2.0.0
pycparser==2.22
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
Quart==0.19.6
requests==2.31.0
schedule==1.2.2
setuptools==70.0.0
simple-websocket==1.0.0
SQLAlchemy==2.0.37
twitchio==2.9.1
typing_extensions==4.11.0
urllib3==2.2.1
webassets==2.0
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
